+++ server/component/style/dataContainer/DataContainerModel.php
@@ -57,7 +57,7 @@ class DataContainerModel extends StyleModel
                     if ($scope) {
                         $scoped_vars = array();
                         foreach ($interpolation_data['data_config_retrieved'] as $key => $value) {
-                            $scoped_vars[$scope . '_' . $key] = $value; // add the scope prefix
+                            $scoped_vars[$scope . '.' . $key] = $value; // add the scope prefix
                         }
                         $entry_record = array_merge($entry_record, $scoped_vars);
                     } else {